requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attr==0.3.2
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Brotli==1.0.9
bs4==0.0.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
chromedriver==2.24.1
clickhouse-driver==0.2.4
cssselect==1.1.0
dataclasses-json==0.5.7
dateparser==1.1.1
demjson3==3.0.5
exceptiongroup==1.0.0rc9
facebook-scraper==0.2.59
facebook-sdk==3.1.0
fake-useragent==0.1.11
frozenlist==1.3.1
h11==0.14.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.12.0
importlib-resources==5.4.0
jsonpath==0.82
lxml==4.9.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
multidict==6.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
numpy==1.23.4
oauthlib==3.2.1
outcome==1.2.0
packaging==21.3
pandas==1.5.1
parse==1.19.0
protobuf==3.20.1
pydantic==1.10.2
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyppeteer-stealth==2.7.3
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-facebook-api==0.14.5
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
rel==0.4.7
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-tor==1.3
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stem==1.8.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing-inspect==0.8.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
w3lib==2.0.1
websocket-client==1.4.1
websockets==10.3
wsproto==1.2.0
yarl==1.8.1
zipp==3.8.1
